+++ src/form/b_textarea.rs
@@ -6,13 +6,11 @@ pub fn BTextarea(
     #[prop(optional)] id: Option<&'static str>,
     #[prop(optional)] name: Option<&'static str>,
     #[prop(optional)] placeholder: Option<&'static str>,
-    #[prop(optional)] value: Option<&'static str>,
+    #[prop(optional, into)] value: MaybeSignal<String>,
     #[prop(attrs, optional)] attributes: Vec<(&'static str, Attribute)>,
 ) -> impl IntoView {
     let mut b_textarea = view! {
-        <textarea class=format!("textarea {}", class.unwrap_or_default()) id=id name=name placeholder=placeholder>
-            {value}
-        </textarea>
+        <textarea class=format!("textarea {}", class.unwrap_or_default()) id=id name=name placeholder=placeholder prop:value=value/>
     };
 
     for (attr_name, attr_value) in attributes {